+++ UnityEditor.TestRunner/TestLaunchers/RemotePlayerTestController.cs
@@ -23,10 +23,14 @@ namespace UnityEditor.TestRunner.TestLaunchers
         [SerializeField]
         private bool m_RegisteredConnectionCallbacks;
 
+        [SerializeField] 
+        private int m_HearbeatTimeOut;
+        
         private IDelayedCallback m_TimeoutCallback;
         
-        public void Init(BuildTarget buildTarget)
+        public void Init(BuildTarget buildTarget, int heartbeatTimeout)
         {
+            m_HearbeatTimeOut = heartbeatTimeout;
             m_PlatformSpecificSetup = new PlatformSpecificSetup(buildTarget);
             m_PlatformSpecificSetup.Setup();
             m_RemoteTestResultReciever = new RemoteTestResultReciever();
@@ -36,8 +40,6 @@ namespace UnityEditor.TestRunner.TestLaunchers
                 EditorConnection.instance.Initialize();
                 DelegateEditorConnectionEvents();
             }
-            
-            m_TimeoutCallback = new DelayedCallback(TimeoutCallback, k_HeartbeatTimeout);
         }
 
         private void DelegateEditorConnectionEvents()
@@ -53,14 +55,14 @@ namespace UnityEditor.TestRunner.TestLaunchers
 
         private void RunStarted(MessageEventArgs messageEventArgs)
         {
-            m_TimeoutCallback.Reset();
+            m_TimeoutCallback?.Reset();
             m_RemoteTestResultReciever.RunStarted(messageEventArgs);
             CallbacksDelegator.instance.RunStartedRemotely(messageEventArgs.data);
         }
 
         private void RunFinished(MessageEventArgs messageEventArgs)
         {
-            m_TimeoutCallback.Clear();
+            m_TimeoutCallback?.Clear();
             m_RemoteTestResultReciever.RunFinished(messageEventArgs);
             m_PlatformSpecificSetup.CleanUp();
 
@@ -69,29 +71,30 @@ namespace UnityEditor.TestRunner.TestLaunchers
 
         private void TestStarted(MessageEventArgs messageEventArgs)
         {
-            m_TimeoutCallback.Reset();
+            m_TimeoutCallback?.Reset();
             CallbacksDelegator.instance.TestStartedRemotely(messageEventArgs.data);
         }
 
         private void TestFinished(MessageEventArgs messageEventArgs)
         {
-            m_TimeoutCallback.Reset();
+            m_TimeoutCallback?.Reset();
             CallbacksDelegator.instance.TestFinishedRemotely(messageEventArgs.data);
         }
         
         private void PlayerAliveHearbeat(MessageEventArgs messageEventArgs)
         {
-            m_TimeoutCallback.Reset();
+            m_TimeoutCallback?.Reset();
         }
 
         private void TimeoutCallback()
         {
-            CallbacksDelegator.instance.RunFailed($"Test execution timed out. No activity received from the player in {k_HeartbeatTimeout/60} minutes.");
+            CallbacksDelegator.instance.RunFailed($"Test execution timed out. No activity received from the player in {m_HearbeatTimeOut} seconds.");
         }
 
         public void PostBuildAction()
         {
             m_PlatformSpecificSetup.PostBuildAction();
+            m_TimeoutCallback = new DelayedCallback(TimeoutCallback, m_HearbeatTimeOut);
         }
 
         public void PostSuccessfulBuildAction()